Animal Header
Alec Austin - I made a header file for our object (Animal).  Feel free to look over it.  It is, of course, subject to change as we need more methods for the assignment, but if there is anything that you wish to add or change as is, feel free to do so.  Just make sure to inform the other two members before committing.