Animal Implementation
Alec Austin - This is the implementation of the Animal class.  Just like the header, it will be subject to change as we need more methods for the assignment, but again if there are any changes you want, feel free to make them.  Just be sure to inform the other members before you commit.